lift SymBuffer, finalize naming
